add Datum.Long to support CDF_TT2000.
svn path=/dasCore/community/autoplot2011/trunk/; revision=11300

+++ dasCoreDatum/src/org/das2/datum/NumberUnits.java
@@ -61,7 +61,7 @@ public class NumberUnits extends Units {
     
     @Override
     public Datum createDatum( long value ) {
-        return new Datum.Double( value, this );
+        return new Datum.Long( value, this );
     }
     
     @Override
